Logical issue resolve ROOM DB and Add Palette fun

+++ app/src/main/java/com/muhammadali/udemy/jetpack/databinding/views/DogListAdapter.kt
@@ -62,7 +62,7 @@ class DogListAdapter(val dogBreedList: ArrayList<DogBreed>) :
 
     override fun onClick(view :View,dogBreed: DogBreed) {
         val action =
-            ListFragmentDirections.actionListFragmentToDetailFragment(dogBreed.id!!.toLong())
+            ListFragmentDirections.actionListFragmentToDetailFragment(dogBreed.id!!)
         Navigation.findNavController(view)
             .navigate(action)
     }
